+++ semcore/data-table/src/components/DataTable/DataTable.tsx
@@ -63,7 +63,7 @@ class DataTableRoot<
     {},
     {},
   typeof DataTableRoot.enhance,
-  { use: DTRow<UniqKeyType>; expandedRows: Set<UniqKeyType>; renderEmptyData: () => React.ReactNode }
+  typeof DataTableRoot.defaultProps
   > {
   static displayName = 'DataTable';
   static style = style;
@@ -81,6 +81,8 @@ class DataTableRoot<
     h: 'fit-content',
     renderEmptyData: () => <NoData py={10} type='nothing-found' description='' w='100%' />,
     variant: 'default',
+    accordionAnimationRows: 40,
+    accordionDuration: 200,
   };
 
   private columns: DTColumn[] = [];
@@ -184,7 +186,7 @@ class DataTableRoot<
     const { totalRows, expandedRows } = this.asProps;
     const flatRows = this.getFlatRows();
 
-    const expandedRowsCount = Array.from(expandedRows ?? []).reduce((acc, rowKey) => {
+    const expandedRowsCount = Array.from(expandedRows ?? []).reduce<number>((acc, rowKey) => {
       const dtRow = flatRows.find((el) => el[UNIQ_ROW_KEY] === rowKey);
       if (dtRow) {
         const expandedRows = dtRow[ACCORDION];
@@ -318,12 +320,14 @@ class DataTableRoot<
       limit,
       variant,
       totalRows,
+      accordionAnimationRows,
     } = this.asProps;
     const { gridTemplateColumns, gridTemplateAreas } = this.gridSettings;
     const { shadowVertical } = this.state;
 
     return {
       accordionDuration,
+      accordionAnimationRows,
       accordionMode,
       columns: this.columns,
       rows: this.getRows(),
@@ -402,7 +406,7 @@ class DataTableRoot<
     }
   });
 
-  handleCellClick = (e: React.SyntheticEvent, opt: { rowIndex: number; colIndex: number; row?: DTRow<UniqKeyType> }) => {
+  handleCellClick = (e: React.SyntheticEvent<HTMLElement>, opt: { rowIndex: number; colIndex: number; row?: DTRow<UniqKeyType> }) => {
     if (lastInteraction.isMouse()) {
       this.initFocusableCell([this.hasFocusableInHeader() ? opt.rowIndex + 1 : opt.rowIndex, opt.colIndex]);
     }
@@ -556,7 +560,7 @@ class DataTableRoot<
         cell.setAttribute('aria-describedby', describedBy);
       }
 
-      cell?.focus();
+      cell?.focus({ focusVisible: true });
 
       if (newRow !== 0) {
         currentHeaderCell?.setAttribute('inert', '');
@@ -758,9 +762,9 @@ class DataTableRoot<
 
       if (cell instanceof HTMLElement) {
         if (hasParent(e.target, cell) && !e.target.dataset.skipTargetFocus) {
-          e.target.focus();
+          e.target.focus({ focusVisible: true });
         } else {
-          cell.focus();
+          cell.focus({ focusVisible: true });
         }
       }