+++ semcore/radio/__tests__/radio.browser-test.tsx
@@ -95,9 +95,6 @@ test.describe('Visual - Radio with group', () => {
     await test.step('Verify keyboard interactions work after mouse', async () => {
       page.locator('[data-ui-name="Value.RadioMark"]').nth(1).click();
       await page.keyboard.press('ArrowDown');
-      await page.keyboard.press('ArrowUp');
-      await page.keyboard.press('ArrowUp');
-      page.locator('label').filter({ hasText: 'Labrador Retriever' }).locator('div').hover();
       await expect(page).toHaveScreenshot();
     });
   });
@@ -368,6 +365,7 @@ test.describe('Functional - Radio with Additional input props', () => {
       const mark = page.locator('[data-ui-name="Value.RadioMark"]').nth(i);
 
       await expect(radio).toHaveAttribute('aria-invalid', 'false');
+      await expect(radio).toHaveAttribute('name', 'radio');
 
       await expect(radio).not.toBeChecked();
 
@@ -398,11 +396,7 @@ test.describe('Functional - Radio with Additional input props', () => {
     });
 
     await test.step('Verify keyboard interactions work after mouse', async () => {
-      if (browserName === 'firefox') {
-        // BUG!
-        await page.keyboard.press('Tab');
-        await page.keyboard.press('Space');
-      } else await page.keyboard.press('ArrowDown');
+      await page.keyboard.press('ArrowDown');
 
       await expect(locators.radioGroup(page)).toHaveAttribute('value', '2');