+++ semcore/data-table/CHANGELOG.md
@@ -2,6 +2,13 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
+## [16.4.2] - 2025-10-30
+
+### Fixed
+
+- Low performance when opening an accordion with a large number of rows.
+- Keyboard interaction after mouse clicking in Safari.
+
 ## [16.4.1] - 2025-10-17
 
 ### Fixed